pyproject.toml
[project]
name = "aiautonomoustraderbot"
version = "0.1.0"
description = "Autonomous trading bot: LangGraph agent on Solana devnet with self-modification sandbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langgraph>=0.2",
    "playwright>=1.57",
    "solana>=0.30",
    "solders>=0.18",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "base58>=2.1",
    "python-binance>=1.0",
    "pandas>=2.0",
    "ta>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["core*", "tools*", "policies*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
